The Lognormal Distribution

Density, distribution function, quantile function, hazard function, cumulative hazard function, and random generation for the Lognormal distribution with parameters shape and scale.

hlnorm(x, meanlog = 0, sdlog = 1, shape = 1 / sdlog, scale = exp(meanlog),
prop = 1, log = FALSE)
Hlnorm(x, meanlog = 0, sdlog = 1, shape = 1 / sdlog, scale = exp(meanlog),
prop = 1, log.p = FALSE)

Arguments

x
vector of quantiles.

meanlog
mean in the Normal distribution.

sdlog, shape
sdlog is standard deviation in the Normal distrimution, shape = 1/sdlog.

scale
is exp(meanlog).

prop
proportionality constant in the extended Lognormal distribution.

log, log.p
logical; if TRUE, probabilities p are given as log(p).

Value

dlnorm gives the density, plnorm gives the distribution function, qlnorm gives the quantile function, hlnorm gives the hazard function, Hlnorm gives the cumulative hazard function, and rlnorm generates random deviates.

Invalid arguments will result in return value NaN, with a warning.

Details

The Lognormal distribution with scale parameter \(a\) and shape parameter \(\sigma\) has hazard function given by $$h(x) = (b/\sigma)(x/\sigma)^(b-1)\exp((x / \sigma)^b)$$ for \(x \ge 0\).