Accelerated Failure Time Regression

The accelerated failure time model with parametric baseline hazard(s). Allows for stratification with different scale and shape in each stratum, and left truncated and right censored data.

aftreg(
  formula = formula(data),
  data = parent.frame(),
  na.action = getOption("na.action"),
  dist = "weibull",
  init,
  shape = 0,
  id,
  param = c("lifeAcc", "lifeExp"),
  control = list(eps = 1e-08, maxiter = 20, trace = FALSE),
  singular.ok = TRUE,
  model = FALSE,
  x = FALSE,
  y = TRUE
)

Arguments

formula
a formula object, with the response on the left of a ~ operator, and the terms on the right. The response must be a survival object as returned by the Surv function.

data
a data.frame in which to interpret the variables named in the formula.

na.action
a missing-data filter function, applied to the model.frame, after any subset argument has been used. Default is options()$na.action.

dist
Which distribution? Default is "weibull", with the alternatives "gompertz", "ev", "loglogistic" and "lognormal". A special case like the exponential can be obtained by choosing "weibull" in combination with shape = 1.

init
vector of initial values of the iteration. Default initial value is zero for all variables.

shape
If positive, the shape parameter is fixed at that value. If zero or negative, the shape parameter is estimated. Stratification is now regarded as a meaningful option even if shape is fixed.

id
If there are more than one spell per individual, it is essential to keep spells together by the id argument. This allows for time-varying covariates.

param
Which parametrization should be used? The lifeAcc uses the parametrization given in the vignette, while the lifeExp uses the same as in the survreg function.

control
a list with components eps (convergence criterion), maxiter (maximum number of iterations), and trace (logical, debug output if TRUE). You can change any component without mention the other(s).

singular.ok
Not used.

model
Not used.

x
Return the design matrix in the model object?

y
Return the response in the model object?

Value

A list of class c("aftreg", "coxreg") with components

coefficients

Fitted parameter estimates.

var

Covariance matrix of the estimates.

loglik

Vector of length two; first component is the value at the initial parameter values, the second componet is the maximized value.

score

The score test statistic (at the initial value).

linear.predictors

The estimated linear predictors.

means

Means of the columns of the design matrix.

w.means

Weighted (against exposure time) means of covariates; weighted relative frequencies of levels of factors.

n

Number of spells in indata (possibly after removal of cases with NA's).

events

Number of events in data.

terms

Used by extractor functions.

assign

Used by extractor functions.

%
wald.test

The Wald test statistic (at the initial value).

y

The Surv vector.

isF

Logical vector indicating the covariates that are factors.

covars

The covariates.

ttr

Total Time at Risk.

levels

List of levels of factors.

formula

The calling formula.

call

The call.

method

The method.

convergence

Did the optimization converge?

fail

Did the optimization fail? (Is NULL if not).

pfixed

TRUE if shape was fixed in the estimation.

param

The parametrization.

Details

The parameterization is different from the one used by survreg, when param = "lifeAcc". The result is then true acceleration of time. Then the model is

$$S(t; a, b, \beta, z) = S_0((t / \exp(b - z\beta))^{\exp(a)})$$ S(t; a, b, beta, z) = S_0((t/exp(b - z beta))^exp(a))

where \(S_0\) is some standardized survivor function. The baseline parameters \(a\) and \(b\) are log shape and log scale, respectively. This is for the default parametrization. With the lifeExp parametrization, some signs are changed: $$b - z beta$$ is changed to $$b + z beta$$. For the Gompertz distribution, the base parametrization is canonical, a necessity for consistency with the shape/scale paradigm (this is new in 2.3).

See also

Examples

data(mort)
aftreg(Surv(enter, exit, event) ~ ses, param = "lifeExp", data = mort)

#> Call:
#> aftreg(formula = Surv(enter, exit, event) ~ ses, data = mort, 
#>     param = "lifeExp")
#> 
#> Covariate          W.mean      Coef Life-Expn  se(Coef)    Wald p
#> ses 
#>            lower    0.416     0         1           (reference)
#>            upper    0.584     0.348     1.416     0.089     0.000 
#> 
#> Baseline parameters:
#> log(scale)                    3.603               0.071     0.000 
#> log(shape)                    0.331               0.058     0.000 
#> Baseline life expectancy:  
#> 
#> Events                    276 
#> Total time at risk         17038 
#> Max. log. likelihood      -1391.3 
#> LR test statistic         16.1 
#> Degrees of freedom        1 
#> Overall p-value           5.91578e-05