Parametric proportional hazards regression

This function is called by aftreg, but it can also be directly called by a user.

aftreg.fit(X, Y, dist, param, strata, offset, init, shape, id, control, pfixed)

Arguments

X
The design (covariate) matrix.

Y
A survival object, the response.

dist
Which baseline distribution?

param
Which parametrization?

strata
A stratum variable.

offset
Offset.

init
Initial regression parameter values.

shape
If positive, a fixed value of the shape parameter in the distribution. Otherwise, the shape is estimated.

id
See corresponding argument to aftreg.

control
Controls convergence and output.

pfixed
A logical indicating fixed shape parameter(s).

Value

coefficients

Estimated regression coefficients plus estimated scale and shape coefficients, sorted by strata, if present.

df

Degrees of freedom; No. of regression parameters.

var

Variance-covariance matrix

loglik

Vector of length 2. The first component is the maximized loglihood with only scale and shape in the model, the second the final maximum.

conver

TRUE if convergence

fail

TRUE if failure

iter

Number of Newton-Raphson iterates.

n.strata

The number of strata in the data.

Details

See aftreg for more detail.

See also