Creates a minimal representation of a data frame.

Given a data frame with a defined response variable, this function creates a unique representation of the covariates in the data frame, vector (matrix) of responses, and a pointer vector, connecting the responses with the corresponding covariates.

cro(dat, response = 1)

Arguments

dat
A data frame

response
The column(s) where the response resides.

Value

A list with components

y

The response.

covar

A data frame with unique rows of covariates.

keys

Pointers from y to covar, connecting each response with its covariate vector.

Details

The rows in the data frame are converted to text strings with paste and compared with match.

Note

This function is based on suggestions by Anne York and Brian Ripley.

See also

Examples

dat <- data.frame(y = c(1.1, 2.3, 0.7), x1 = c(1, 0, 1), x2 = c(0, 1, 0))
cro(dat)

#> $y
#> [1] 1.1 2.3 0.7
#> 
#> $covar
#>   x1 x2
#> 1  1  0
#> 2  0  1
#> 
#> $keys
#> [1] 1 2 1
#>