Infant mortality and maternal death, Sweeden 1821--1894.

Matched data on infant mortality, from seven parishes in Sweden, 1821--1894.

data(infants)

Format

A data frame with 80 rows and five variables.

stratum

Triplet No. Each triplet consist of one infant whose mother died (a case), and two controls, i.e, infants whose mother did not die. Matched on covariates below.

enter

Age (in days) of case when its mother died.

exit

Age (in days) at death or right censoring (at age 365 days).

event

Follow-up ends with death (1) or right censoring (0).

mother

dead for cases, alive for controls.

age

Mother's age at infant's birth.

sex

The infant's sex.

parish

Birth parish, either Nedertornea or not Nedertornea.

civst

Civil status of mother, married or unmarried.

ses

Socio-economic status of mother, either farmer or not farmer.

year

Year of birth of the infant.

Details

From 5641 first-born in seven Swedish parishes 1820-1895, from Fleninge in the very south to Nedertorneå in the very north, those whose mother died during their first year of life were selected, in all 35 infants. To each of them, two cotrols were selected by matching on the given covariates.

Source

Data originate from The Demographic Data Base, Umeå University, Umeå, Sweden, http://www.ddb.umu.se.

References

Broström, G. (1987). The influence of mother's death on infant mortality: A case study in matched data survival analysis. Scandinavian Journal of Statistics 14, 113-123.

Examples

data(infants)
fit <- coxreg(Surv(enter, exit, event) ~ strata(stratum) + mother, data
= infants)
fit

#> Call:
#> coxreg(formula = Surv(enter, exit, event) ~ strata(stratum) + 
#>     mother, data = infants)
#> 
#> Covariate           Mean       Coef     Rel.Risk   S.E.    Wald p
#> mother 
#>            alive    0.763     0         1 (reference)
#>             dead    0.237     2.605    13.534     0.757     0.001 
#> 
#> Events                    21 
#> Total time at risk         21616 
#> Max. log. likelihood      -10.815 
#> LR test statistic         19.65 
#> Degrees of freedom        1 
#> Overall p-value           9.31744e-06

fit.w <- phreg(Surv(enter, exit, event) ~ mother + parish + ses, data =
infants)
fit.w ## Weibull proportional hazards model.

#> Call:
#> phreg(formula = Surv(enter, exit, event) ~ mother + parish + 
#>     ses, data = infants)
#> 
#> Covariate          W.mean      Coef Exp(Coef)  se(Coef)    Wald p
#> mother 
#>            alive    0.763     0         1           (reference)
#>             dead    0.237     2.262     9.601     0.507     0.000 
#> parish 
#>            other    0.723     0         1           (reference)
#>      Nedertornea    0.277     2.052     7.786     0.498     0.000 
#> ses 
#>            other    0.806     0         1           (reference)
#>           farmer    0.194    -0.346     0.708     0.665     0.603 
#> 
#> log(scale)                   13.275               3.372     0.000 
#> log(shape)                   -0.979               0.533     0.066 
#> 
#> Events                    21 
#> Total time at risk         21616 
#> Max. log. likelihood      -139.17 
#> LR test statistic         33.46 
#> Degrees of freedom        3 
#> Overall p-value           2.57e-07