Rye prices, Scania, southern Sweden, 1801-1894.

The data consists of yearly rye prices from 1801 to 1894. Logged and detrended, so the time series is supposed to measure short term fluctuations in rye prices.

data(scania)

Format

A data frame with 94 observations in two columns on the following 2 variables.

year

The year the price is recorded.

foodprices

Detrended log rye prices.

Details

The Scanian area in southern Sweden was during the 19th century a mainly rural area.

Source

The Scanian Economic Demographic Database.

References

Jörberg, L. (1972). A History of Prices in Sweden 1732-1914, CWK Gleerup, Lund.

Examples

data(logrye)
summary(logrye)

#>       year        foodprices       
#>  Min.   :1801   Min.   :-0.378000  
#>  1st Qu.:1824   1st Qu.:-0.124250  
#>  Median :1848   Median :-0.004500  
#>  Mean   :1848   Mean   : 0.002457  
#>  3rd Qu.:1871   3rd Qu.: 0.125500  
#>  Max.   :1894   Max.   : 0.423000