Male mortality in ages 40-60, nineteenth century

Males born in the years 1800-1820 and surving at least 40 years in the parish Skellefteå in northern Sweden are followed from their fortieth birthday until death or the sixtieth birthday, whichever comes first.

data(mort)

Format

A data frame with 2058 observations on the following 6 variables.

id

Personal identification number.

enter

Start of duration. Measured in years since the fortieth birthday.

exit

End of duration. Measured in years since the fortieth birthday.

event

a logical vector indicating death at end of interval.

birthdate

The birthdate in decimal form.

ses

Socio-economic status, a factor with levels lower, upper

Note

This data set is also known, and accessible, as male.mortality

Details

The interesting explanatory covariate is ses (socioeconomic status), which is a time-varying covariate. This explains why several individuals are representated by more than one record each. Left trucation and right censoring are introduced this way.

Source

Data is coming from The Demographic Data Base, Umea University, Umeå, Sweden.

References

http://www.ddb.umu.se

Examples

data(mort)
coxreg(Surv(enter, exit, event) ~ ses, data = mort)

#> Call:
#> coxreg(formula = Surv(enter, exit, event) ~ ses, data = mort)
#> 
#> Covariate           Mean       Coef     Rel.Risk   S.E.    Wald p
#> ses 
#>            lower    0.416     0         1 (reference)
#>            upper    0.584    -0.480     0.619     0.121     0.000 
#> 
#> Events                    276 
#> Total time at risk         17038 
#> Max. log. likelihood      -1845.3 
#> LR test statistic         15.81 
#> Degrees of freedom        1 
#> Overall p-value           7.01379e-05