Old age mortality, Sundsvall, Sweden, 1860-1880.

The data consists of old age life histories from 1 January 1860 to 31 december 1880, 21 years. Only (parts of) life histories above age 60 is considered.

data(oldmort)

Format

A data frame with 6508 observations from 4603 persons on the following 13 variables.

id

Identification number.

enter

Start age for the interval.

exit

Stop age for the interval.

event

Indicator of death; equals TRUE if the person died at the end of the interval, FALSE otherwise.

birthdate

Birthdate as a real number (i.e., "1765-06-27" is 1765.490).

m.id

Mother's identification number.

f.id

Father's identification number.

sex

Gender, a factor with levels male female

civ

Civil status, a factor with levels unmarried married widow

ses.50

Socio-economic status at age 50, a factor with levels middle unknown upper farmer lower

birthplace

a factor with levels parish region remote

imr.birth

Infant mortality rate at birth in the region of birth

region

Subregion of Sundsvall, a factor with levels town industry rural

Details

The Sundsvall area in mid-Sweden was during the 19th century a fast growing forest industry. At the end of the century, it was one of the largest sawmill area in Europe. The town Sundsvall is fast growing part of the region and center for the commerse.

Source

The Demographic Data Base, Umeå University, Sweden.

References

Edvinsson, S. (2000). The Demographic Data Base at Umeå University: A resource for historical studies. In Hall, McKaa, and Thorvaldsen (eds), "Handbook of International Historical Microdata for Population Research", Minnesota Population Center, Minneapolis.

Examples

data(oldmort)
summary(oldmort)

#>        id                enter            exit          event        
#>  Min.   :765000603   Min.   :60.00   Min.   : 60.00   Mode :logical  
#>  1st Qu.:797001170   1st Qu.:60.00   1st Qu.: 63.88   FALSE:4524     
#>  Median :804001545   Median :60.07   Median : 68.51   TRUE :1971     
#>  Mean   :803652514   Mean   :64.07   Mean   : 69.89                  
#>  3rd Qu.:812001564   3rd Qu.:66.88   3rd Qu.: 74.73                  
#>  Max.   :826002672   Max.   :94.51   Max.   :100.00                  
#>                                                                      
#>    birthdate         m.id                f.id               sex      
#>  Min.   :1765   Min.   :     6039   Min.   :     2458   male  :2884  
#>  1st Qu.:1797   1st Qu.:766000610   1st Qu.:763000610   female:3611  
#>  Median :1805   Median :775000742   Median :772000649                
#>  Mean   :1804   Mean   :771271398   Mean   :762726961                
#>  3rd Qu.:1812   3rd Qu.:783000743   3rd Qu.:780001077                
#>  Max.   :1820   Max.   :802000669   Max.   :797001468                
#>                 NA's   :3155        NA's   :3310                     
#>         civ           ses.50      birthplace     imr.birth           region    
#>  unmarried: 557   middle : 233   parish:3598   Min.   : 4.348   town    : 657  
#>  married  :3638   unknown:2565   region:1503   1st Qu.:12.709   industry:2214  
#>  widow    :2300   upper  :  55   remote:1394   Median :14.234   rural   :3624  
#>                   farmer :1562                 Mean   :15.209                  
#>                   lower  :2080                 3rd Qu.:17.718                  
#>                                                Max.   :31.967                  
#>                                                                                

## maybe str(oldmort) ; plot(oldmort) ...