Old age mortality, Scania, southern Sweden, 1813-1894.

The data consists of old age life histories from 1 January 1813 to 31 december 1894. Only (parts of) life histories above age 50 is considered.

data(scania)

Format

A data frame with 1931 observations from 1931 persons on the following 9 variables.

id

Identification number (enumeration).

enter

Start age for the interval.

exit

Stop age for the interval.

event

Indicator of death; equals TRUE if the person died at the end of the interval, FALSE otherwise.

birthdate

Birthdate as a real number (i.e., "1765-06-27" is 1765.490).

sex

Gender, a factor with levels male female.

parish

One of five parishes in Scania, coded 1, 2, 3, 4, 5. Factor.

ses

Socio-economic status at age 50, a factor with levels upper and lower.

immigrant

a factor with levels no region and yes.

Details

The Scanian area in southern Sweden was during the 19th century a mainly rural area.

Source

The Scanian Economic Demographic Database, Lund University, Sweden.

References

http://www.ed.lu.se/databases

Examples

data(scania)
summary(scania)

#>        id             enter         exit           event          birthdate   
#>  Min.   :   1.0   Min.   :50   Min.   :50.00   Min.   :0.0000   Min.   :1763  
#>  1st Qu.: 483.5   1st Qu.:50   1st Qu.:55.37   1st Qu.:0.0000   1st Qu.:1786  
#>  Median : 966.0   Median :50   Median :62.70   Median :1.0000   Median :1808  
#>  Mean   : 966.0   Mean   :50   Mean   :63.97   Mean   :0.5624   Mean   :1806  
#>  3rd Qu.:1448.5   3rd Qu.:50   3rd Qu.:72.05   3rd Qu.:1.0000   3rd Qu.:1825  
#>  Max.   :1931.0   Max.   :50   Max.   :80.00   Max.   :1.0000   Max.   :1845  
#>      sex      parish      ses       immigrant 
#>  male  :975   1: 159   upper: 375   no : 759  
#>  female:956   2: 181   lower:1556   yes:1172  
#>               3: 180                          
#>               4: 258                          
#>               5:1153                          
#>