Swedish population and deaths in ages 61--80, 2007.

The Swedish population and No. of deaths by age and sex in the ages 61--80. Data from the year 2007.

data(swe07)

Format

A data frame with 80 rows and five variables.

pop

Average population size during the year 2007 by age and sex.

deaths

Number of deaths by age and sex during the year 2007.

sex

Sex.

age

Age.

log.pop

The logarithm of the first variable, pop. Included for convenience, may be used as an offset in a Poisson regression.

Details

The average population is calculated as the mean of the population 1 January 2007 and 1 January 2008.

Source

Data is taken from Statistics Sweden.

References

http://www.scb.se

Examples

data(swe07)
fit <- glm(deaths ~ offset(log.pop) + sex * as.factor(age), family = poisson, data = swe07)
drop1(fit, test = "Chisq")  ## Proportional hazards?

#> Single term deletions
#> 
#> Model:
#> deaths ~ offset(log.pop) + sex * as.factor(age)
#>                    Df Deviance    AIC    LRT Pr(>Chi)
#> <none>                   0.000 410.58                
#> sex:as.factor(age) 19   10.232 382.81 10.232    0.947